feat(导航): 添加Scrcpy高级设置页面路由和功能
- 在Routes.kt中添加ScrcpyAdvanced路由
- 在MainActivity中注册ScrcpyAdvancedScreen
- 重构Scrcpy设备页面，支持打开高级设置
- 新增Scrcpy高级设置页面组件
